+++ android-sdk-ui/src/main/java/com/braze/ui/actions/UriAction.kt
@@ -9,6 +9,7 @@ import android.net.Uri
 import android.os.Build
 import android.os.Bundle
 import androidx.annotation.VisibleForTesting
+import com.braze.BrazeInternal
 import com.braze.Constants
 import com.braze.enums.Channel
 import com.braze.IBrazeDeeplinkHandler
@@ -130,7 +131,7 @@ open class UriAction : IAction {
      * @see [UriAction.getIntentArrayWithConfiguredBackStack]
      */
     protected fun openUriWithWebViewActivityFromPush(context: Context, uri: Uri, extras: Bundle?) {
-        val configurationProvider = BrazeConfigurationProvider(context)
+        val configurationProvider = BrazeInternal.getConfigurationProvider(context)
         try {
             val webViewIntent = getWebViewActivityIntent(context, uri, extras)
             context.startActivities(
@@ -153,7 +154,7 @@ open class UriAction : IAction {
      * @see [UriAction.getIntentArrayWithConfiguredBackStack]
      */
     protected fun openUriWithActionViewFromPush(context: Context, uri: Uri, extras: Bundle?) {
-        val configurationProvider = BrazeConfigurationProvider(context)
+        val configurationProvider = BrazeInternal.getConfigurationProvider(context)
         try {
             val uriIntent = getActionViewIntent(context, uri, extras)
             context.startActivities(
@@ -173,7 +174,7 @@ open class UriAction : IAction {
      * Returns an intent that opens the uri inside of a [BrazeWebViewActivity].
      */
     protected fun getWebViewActivityIntent(context: Context, uri: Uri, extras: Bundle?): Intent {
-        val configurationProvider = BrazeConfigurationProvider(context)
+        val configurationProvider = BrazeInternal.getConfigurationProvider(context)
         val customWebViewActivityClassName = configurationProvider.customHtmlWebViewActivityClassName
 
         // If the class is valid and is manifest registered, use it as the launching intent